backend/requirements.txt
Flask
python-dotenv
psycopg2-binary
SQLAlchemy
Flask-CORS
Werkzurg
PyJWT
